pyproject.toml
[project]
name = "grocery-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.70.0",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
]
